+++ runtime/include/pct_strategy.h
@@ -11,24 +11,23 @@
 // Although it's impossible to predict the exact number of switches(since it's
 // equivalent to the halt problem), k should be good approximation
 template <typename TargetObj>
-struct PctStrategy : Strategy {
+struct PctStrategy : public BaseStrategyWithThreads<TargetObj> {
   // TODO: doc about is_another_required
   explicit PctStrategy(size_t threads_count,
-                       const std::vector<TaskBuilder>& constructors,
+                       std::vector<TaskBuilder> constructors,
                        bool is_another_required)
       : threads_count(threads_count),
         current_depth(1),
         current_schedule_length(0),
-        constructors(constructors),
-        threads(),
         is_another_required(is_another_required) {
-    round_schedule.resize(threads_count, -1);
+    this->constructors = std::move(constructors);
+    Strategy::round_schedule.resize(threads_count, -1);
 
     std::random_device dev;
     rng = std::mt19937(dev());
-    constructors_distribution =
+    this->constructors_distribution =
         std::uniform_int_distribution<std::mt19937::result_type>(
-            0, constructors.size() - 1);
+            0, this->constructors.size() - 1);
 
     // We have information about potential number of resumes
     // but because of the implementation, it's only available in the task.
@@ -43,19 +42,20 @@ struct PctStrategy : Strategy {
     //      avg_k += task.GetSuspensionPoints();
     //      task.Terminate();
     //    }
-    avg_k = avg_k / constructors.size();
+    avg_k = avg_k / this->constructors.size();
 
     PrepareForDepth(current_depth, avg_k);
 
     // Create queues.
     for (size_t i = 0; i < threads_count; ++i) {
-      threads.emplace_back();
+      this->threads.emplace_back();
     }
   }
 
   // If there aren't any non returned tasks and the amount of finished tasks
   // is equal to the max_tasks the finished task will be returned
   std::tuple<Task&, bool, int> Next() override {
+    auto& threads = this->threads;
     size_t max = std::numeric_limits<size_t>::min();
     size_t index_of_max = 0;
     // Have to ignore waiting threads, so can't do it faster than O(n)
@@ -101,14 +101,14 @@ struct PctStrategy : Strategy {
 
     if (threads[index_of_max].empty() ||
         threads[index_of_max].back()->IsReturned()) {
-      auto constructor = constructors.at(constructors_distribution(rng));
+      auto constructor = this->constructors.at(this->constructors_distribution(rng));
       if (is_another_required) {
         auto names = CountNames(index_of_max);
         // TODO: выглядит непонятно и так себе
         while (true) {
           names.insert(constructor.GetName());
           if (names.size() == 1) {
-            constructor = constructors.at(constructors_distribution(rng));
+            constructor = this->constructors.at(this->constructors_distribution(rng));
           } else {
             break;
           }
@@ -116,7 +116,7 @@ struct PctStrategy : Strategy {
       }
 
       threads[index_of_max].emplace_back(
-          constructor.Build(&state, index_of_max, new_task_id++));
+          constructor.Build(&this->state, index_of_max, Strategy::new_task_id++));
       return {threads[index_of_max].back(), true, index_of_max};
     }
 
@@ -127,39 +127,20 @@ struct PctStrategy : Strategy {
     assert(false && "unimplemented");
   }
 
-  std::optional<std::tuple<Task&, int>> GetTask(int task_id) override {
-    // TODO: can this be optimized?
-    int thread_id = 0;
-    for (auto& thread : threads) {
-      size_t tasks = thread.size();
-
-      for (size_t i = 0; i < tasks; ++i) {
-        Task& task = thread[i];
-        if (task->GetId() == task_id) {
-          std::tuple<Task&, int> result = { task, thread_id };
-          return result;
-        }
-      }
-
-      thread_id++;
-    }
-    return std::nullopt;
-  }
-
   void StartNextRound() override {
-    new_task_id = 0;
+    Strategy::new_task_id = 0;
 
     //    log() << "depth: " << current_depth << "\n";
     // Reconstruct target as we start from the beginning.
-    TerminateTasks();
-    for (auto& thread : threads) {
+    this->TerminateTasks();
+    for (auto& thread : this->threads) {
       // We don't have to keep references alive
       while (thread.size() > 0) {
         thread.pop_back();
       }
     }
 
-    state.Reset();
+    this->state.Reset();
     // Update statistics
     current_depth++;
     if (current_depth >= 50) {
@@ -174,65 +155,21 @@ struct PctStrategy : Strategy {
     log() << "k: " << new_k << "\n";
     PrepareForDepth(current_depth, new_k);
 
-    for (auto& thread : threads) {
+    for (auto& thread : this->threads) {
       thread = StableVector<Task>();
     }
   }
 
-  void ResetCurrentRound() override {
-    TerminateTasks();
-    state.Reset();
-
-    for (auto& thread : threads) {
-      size_t tasks_in_thread = thread.size();
-      for (size_t i = 0; i < tasks_in_thread; ++i) {
-        if (!thread[i]->IsRemoved()) {
-          thread[i] = thread[i]->Restart(&state);
-        }
-      }
-    }
-  }
-
-  int GetValidTasksCount() const override {
-    int non_removed_tasks = 0;
-    for (auto& thread : threads) {
-      for (size_t i = 0; i < thread.size(); ++i) {
-        auto& task = thread[i];
-        if (!task.get()->IsRemoved()) {
-          non_removed_tasks++;
-        }
-      }
-    }
-    return non_removed_tasks;
-  }
-
-  ~PctStrategy() { TerminateTasks(); }
-
-protected:
-  int GetNextTaskInThread(int thread_index) const override {
-    auto& thread = threads[thread_index];
-    int task_index = round_schedule[thread_index];
-
-    while (
-      task_index < static_cast<int>(thread.size()) &&
-      (
-        task_index == -1 ||
-        thread[task_index].get()->IsReturned() ||
-        thread[task_index].get()->IsRemoved()
-      )
-    ) {
-      task_index++;
-    }
-
-    return task_index;
+  ~PctStrategy() {
+    this->TerminateTasks();
   }
 
 private:
   std::unordered_set<std::string> CountNames(size_t except_thread) {
     std::unordered_set<std::string> names;
 
-    for (size_t i = 0; i < threads.size(); ++i) {
-      auto& thread = threads[i];
+    for (size_t i = 0; i < this->threads.size(); ++i) {
+      auto& thread = this->threads[i];
       if (thread.empty() || i == except_thread) {
         continue;
       }
@@ -261,34 +198,12 @@ private:
     }
   }
 
-  void TerminateTasks() {
-    assert(round_schedule.size() == threads.size() && "sizes expected to be the same");
-    round_schedule.assign(round_schedule.size(), -1);
-
-    for (auto& thread : threads) {
-      for (size_t i = 0; i < thread.size(); ++i) {
-        if (!thread[i]->IsReturned()) {
-          thread[i]->Terminate();
-        }
-      }
-    }
-  }
-
-  TargetObj state{};
-  std::vector<TaskBuilder> constructors;
   std::vector<size_t> k_statistics;
   size_t threads_count;
   size_t current_depth;
   size_t current_schedule_length;
   std::vector<size_t> priorities;
   std::vector<size_t> priority_change_points;
-  // RoundRobinStrategy struct is the owner of all tasks, and all
-  // references can't be invalidated before the end of the round,
-  // so we have to contains all tasks in queues(queue doesn't invalidate the
-  // references)
-  std::vector<StableVector<Task>> threads;
   bool is_another_required;
-  std::uniform_int_distribution<std::mt19937::result_type>
-      constructors_distribution;
   std::mt19937 rng;
 };
